refactor(a11y-close-contrast): Splititng container styles out into their own files
