+++ src/checkout/template/containerStyle/index.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export { getContainerStyle } from './base';
