+++ src/checkout/template/sandboxStyle/index.js
@@ -1,0 +1,3 @@
+/* @flow */
+
+export { getSandboxStyle } from './base';
